Refactor IdV Confirmations & Questions controller (#790)
**Why**:
- To better track IdV analytics. Previously, we were only tracking
failures in the KBV scenario.

**How**:
- Add an analytics call to ReviewController to track success and
failure for the initial IdV submission (prior to KBV and/or phone
confirmation)
- Move most of the code in ConfirmationsController to
QuestionsController since it only applies when KBV is on
- Check if KBV is on in QuestionsController#index. If not, redirect
directly to ConfirmationsController. Previously, we were checking if
questions were available even when KBV was turned off.
- Add an analytics call in QuestionsController to track success and
failure when KBV is on
- Only call analytics in ConfirmationsController if KBV is off to
prevent duplicate events when KBV is on.