User can only view recovery code 1x during sign up (#848)
**Why**: Re-visiting the page creates a new recovery code, which is
potentially confusing. Once someone has seen the recovery code during
the sign up flow, they will be redirected to the profile page if the pae
refreshes for any reason.

Note that this is *not* the behavior for the regular "create recovery
code" page that someone accesses after signup. In order to create
different behavior, I separated the two views into separate controllers.