Refactor Resend Email Confirmation flow (#924)
**Why**:
- The `/sign_up/enter_email/resend` route was being visited directly in
some feature specs, but there weren't any other specs that verified
that the user ended up on that endpoint, which made it look like the
endpoint wasn't being used. Technically, it wasn't being used since
the view contents were being rendered directly in
`EmailConfirmationsController#process_unsuccessful_confirmation`

- Some tests were duplicated between controller and feature specs

**How**:
- Instead of rendering the `/sign_up/email_resend/new` view when a user
tries to confirm their email with an invalid or expired token, we
redirect to the `/sign_up/enter_email/resend` endpoint. That way, the
URL changes to reflect the contents of the page. This then allows us
to have tests that visit this endpoint directly since we've tested how
a user can end up there.

- Remove feature specs that were already covered by controller tests
- Replace other feature specs with controller and view specs