Refactor code related to PhoneController#create
**Why**: Same reason as the Finance-related refactor, so I'll copy
and paste the same commit message, with some additional comments.

- Make it more readable and easier to understand
- Reduce the cognitive load on the developer who is trying to read the
  code for the first time
- Adhere to our controller design principles
- Adhere to Single Responsibility Principle
- Make it easier to test

**How**:
- Refactor `PhoneStep` class such that the only public method
called is `#submit`, and have it return a `FormResponse` object that
the controller can use to determine how to proceed, and to capture
analytics for this event.
- Remove the idv_session manipulation from the vendor validation class
and from the private methods of the `PhoneStep` class, and
into the `submit` method.
- Write the `submit` method in an easy-to-follow narrative that shows
the sequence of events, and consequences of the result of the operation
all in the same place. In this case, the operation is the form
submission, and we must ask whether or not the submission was
successful. We determine if the submission was successful by asking if
both the form validations and the vendor validations passed. If it was
successful, then we update the idv_session and set the
phone_confirmation to `true`, otherwise, we set the confirmation
to false, then we return the result in the form of a `FormResponse`
object that responds to `success?` and `errors`, and where the errors
are the combined errors from the form and vendor validations.
All we need to know about this operation and its side effects are in
this one method, whereas before, you had to jump back and forth between
multiple files to understand the entire operation. Now, we don't need
to look elsewhere to understand the flow of events. If we want to dig
deeper, then we can look into the classes that handle the form or vendor
validations, and see how they operate, but it's not necessary to
understand what is happening. We trust that they are doing the right
thing when we ask them if they are valid or not.
- Add an `#errors` method to the vendor validator and update
`PhoneStep#vendor_errors` to return those errors as opposed to
fetching them from the idv_session, where they may or may not exist.
- Update the way a controller can determine if the Phone step was
completed by setting `idv_session.phone_confirmation` explicity to
`true` or `false` to adhere to the "Tell, Don't Ask" principle.