Replace browserify-rails with rails-webpacker
**Why**: rails-webpacker compiles the assets much faster than
browserify-rails. With browserify-rails, our assets take so long to
compile that the initial requests to the idp time out locally. I have
not observed that with webpacker.

Additionally, rails-webpacker is recommended by the maintainer of
browserify-rails for projects that have already made the step towards
using commonjs modules.

One tradeoff of this is that rails-webpacker expects yarn, so we'll need
to move from npm to yarn for this to work.

Another tradeoff is that teaspoon does not appear to play nicely with
rails-webpacker, so I removed it and replaced it with a bare-metal
mocha.